Restructuring WriteQuery module
