+++ server/prisma-rs/query-engine/core/src/executor.rs
@@ -1,0 +1,9 @@
+//! A slightly more generic interface over executing read and write queries
+
+use crate::{WriteQuery, WriteQueryResult};
+
+pub struct Executor;
+
+impl Executor {
+
+}
