+++ server/prisma-rs/query-engine/core/src/lib.rs
@@ -6,6 +6,7 @@ mod query_ast;
 mod query_results;
 mod read_query_executor;
 mod mutations;
+mod executor;
 
 pub mod ir;
