+++ server/prisma-rs/query-engine/prisma/src/req_handlers/graphql.rs
@@ -3,6 +3,7 @@ use crate::{context::PrismaContext, data_model::Validatable, error::PrismaError,
 use core::{
     ir::{self, Builder},
     Query, RootBuilder,
+    CoreResult,
 };
 use connector::mutaction::DatabaseMutactionResult;
 use graphql_parser as gql;
@@ -66,14 +67,25 @@ fn handle_safely(req: PrismaRequest<GraphQlBody>, ctx: &PrismaContext) -> Prisma
     let mut queries = rb.build()?;
 
     // Execute mutations first!
-    let writes: CoreError<Vec<DatabaseMutactionResult>> = queries
-        .iter_mut()
-        .filter_map(|q| match q {
-            Query::Write(q) => Some(q),
-            _ => None,
-        })
-        .map(|wq| ctx.write_query_executor.execute(wq))
-        .collect();
+    let (writes, reads) = queries
+        .into_iter()
+        .fold((vec![], vec![]), |(mut first, mut second), query| {
+            match query {
+                Query::Write(WriteQuery {
+                    root, nested: _,
+                }) => {
+                    let write_results = ctx.write_query_executor.execute(root);
+
+
+                    first.push(ctx.write_query_executor.execute(wq.root));
+
+
+                },
+                Query::Read(rq) => second.push(rq),
+            };
+
+            (first, second)
+        });
 
     // let ir = match queries {
     //     Ok(queries) => match dbg!(ctx.read_query_executor.execute(&queries)) {
